Moved favicon setting from middleware to route-guards. - Removed unauthenticated middleware - Moved some route guards to the same file
