+++ shell/components/templates/unauthenticated.vue
@@ -5,7 +5,6 @@ import FixedBanner from '@shell/components/FixedBanner';
 export default {
   mixins:     [Brand],
   components: { FixedBanner },
-  middleware: ['unauthenticated'],
 };
 </script>
